Update rounds duration logic
Instead of requiring the total score to equal the total rounds, I
updated the logic so that the game is over once one player scores more
than half of the points available. This way, if someone plays (5) rounds
and wins 3, the game would be over even though the total score is less
than the total number of rounds.